requirements.txt
python-telegram-bot==21.0.1
Flask==3.0.3
SQLAlchemy==2.0.25
psycopg2-binary==2.9.9
python-dotenv==1.0.1
APScheduler==3.10.4
cryptography==42.0.4
requests==2.31.0
pyotp==2.9.0
steam==1.4.4
gevent==23.9.1
yookassa==3.3.0
lxml==4.9.3 # Добавлено для FunPayCardinal
steam==1.4.4
eventemitter==0.2.0
protobuf==3.20.3 # Или конкретная версия, совместимая с вашей версией steam
# googleapis-common-protos # Может потребоваться, попробуйте сначала без нее